libs/flock_resources/pyproject.toml
[tool.poetry]
name = "flock-resources"
version = "0.0.1"
description = ""
authors = ["Adam Yodinsky <27074934+adamyodinsky@users.noreply.github.com>"]
readme = "README.md"
packages = [
    { include = "flock_builder", from = "src" },
    { include = "flock_resources", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.7"
pyyaml = "^6.0"
langchain = "^0.0.259"
chromadb = "^0.3.21"
tiktoken = "^0.3.3"
google-search-results = "^2.4.2"
openai = "^0.27.4"
faiss-cpu = "^1.7.3"
odmantic = "^0.9.2"
wolframalpha = "^5.0.0"
pdfminer-six = "^20221105"
azure-core = "^1.26.4"
beautifulsoup4 = "^4.12.2"
pytest-playwright = "^0.3.3"
flock-resource-store = {path = "../flock_resource_store", develop = true}
flock-schemas = {path = "../flock_schemas", develop = true}
langchain-experimental = "^0.0.8"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
python-dotenv = "^1.0.0"
flock-common = {path = "../../libs/flock_common", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
